Split directories api part
